+++ src-tauri/src/game/mod.rs
@@ -30,8 +30,7 @@ const GITHUB_API_URL: &str =
 const GAMES_JSON_URL: &str =
     "https://raw.githubusercontent.com/Kesomannen/gale/refs/heads/master/src-tauri/games.json";
 
-const BUNDLED_GAMES_JSON: &str =
-    include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/", "games.json"));
+const BUNDLED_GAMES_JSON: &str = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/games.json"));
 
 const BUILD_TIME: &str = env!("BUILD_TIME");
 
@@ -75,6 +74,11 @@ fn get_cached_games() -> Result<GamesCache<'static>> {
 }
 
 pub async fn update_list_task(app: &AppHandle) -> Result<()> {
+    if cfg!(debug_assertions) {
+        info!("skipping games list update in debug mode");
+        return Ok(());
+    }
+
     let str = app
         .http()
         .get(GAMES_JSON_URL)
@@ -92,7 +96,7 @@ pub async fn update_list_task(app: &AppHandle) -> Result<()> {
     });
 
     let cache = GamesCache {
-        date: date.clone(),
+        date,
         games,
     };
 
@@ -128,8 +132,7 @@ async fn get_last_commit_date(app: &AppHandle) -> Result<DateTime<Utc>> {
         .error_for_status()?
         .json()
         .await?;
-    let date = response
-        .get(0)
+    let date = response.first()
         .ok_or_eyre("github api response contained no entries")?
         .commit
         .author
@@ -149,7 +152,7 @@ pub fn from_slug(slug: &str) -> Option<Game> {
 }
 
 pub fn last_updated() -> DateTime<Utc> {
-    return GAMES.0;
+    GAMES.0
 }
 
 #[derive(Deserialize, Debug)]