+++ src-tauri/src/profile/install/queue.rs
@@ -363,7 +363,7 @@ fn rollback_batch(batch: &InstallBatch, app: &AppHandle, count: usize) -> Result
                     });
             }
 
-            profile.save(&app, true)?;
+            profile.save(app, true)?;
             Ok(())
         }
     }
@@ -418,7 +418,7 @@ fn try_cache_install(batch: &InstallBatch, index: usize, app: &AppHandle) -> Res
 
     install.clone().insert_into(profile)?;
 
-    profile.save(&app, true)?;
+    profile.save(app, true)?;
 
     emit(
         InstallEvent::AddProgress {
@@ -550,7 +550,7 @@ fn install_from_download(
     installer.install(&cache_path, package_name, profile)?;
     install.clone().insert_into(profile)?;
 
-    profile.save(&app, true)?;
+    profile.save(app, true)?;
 
     emit(InstallEvent::AddProgress { mods: 1, bytes: 0 }, app);