+++ src/routes/+layout.svelte
@@ -18,6 +18,7 @@
 	import { listen, type UnlistenFn } from '@tauri-apps/api/event';
 	import type { ProfileInfo, ManagedGameInfo } from '$lib/types';
 	import { refreshLanguage } from '$lib/i18n';
+	import MissingProfilesDialog from '$lib/components/dialogs/MissingProfilesDialog.svelte';
 
 	type Props = {
 		children?: Snippet;
@@ -80,4 +81,5 @@
 
 	<InstallModDialog />
 	<WelcomeDialog />
+	<MissingProfilesDialog />
 </Tooltip.Provider>