+++ CHANGELOG.md
@@ -2,9 +2,21 @@
 
 ## Unreleased
 
+### Added
+
+- Missing profiles dialog for when gale detects profiles that have been manually moved or deleted
+  - The previous behaviour was to automatically wipe these profiles from the database
+  - You now get the option, per profile, to either locate the profile or have gale wipe it
+
+### Changed
+
+- Tweaked verbosity level to hide excessive debug logs
+
 ### Fixed
 
 - Typo in notification
+- Required shimloader files being put in the wrong directory
+- Shimloader .pak being put in the `mod` directory rather than `pak`
 
 ## 1.12.0 (2025-02-05)